Refactored messaging and contacts packages into modules
