--- plebnet/contacts/__init__.py
@@ -1,2 +1,0 @@
-from plebnet.contacts.AddressBook import AddressBook
-from plebnet.contacts.Contact import Contact
