+++ plebnet/messaging.py
@@ -1,11 +1,43 @@
+import pickle
 import socket
 import collections
 import threading
 import time
-import pickle
 
-from plebnet.messaging.MessageConsumer import MessageConsumer
+from abc import ABC
+from interface import implements, Interface
 
+class MessageSender:
+    
+    def __init__(self, host: str, port: int):
+
+        self.port = port
+        self.host = host
+        
+
+    def sendMessage(self, data):
+ 
+        messageBody = pickle.dumps(data)
+
+        s = socket.socket()   
+               
+        s.connect((self.host, self.port)) 
+
+        s.send(str(len(messageBody)).encode('utf-8'))
+
+        s.recv(1)
+
+        s.send(messageBody)
+
+        s.close()
+
+
+class MessageConsumer(Interface):
+    
+    def notify(self, message):
+        
+        pass
+    
 
 class MessageReceiver:
     """
@@ -76,3 +108,47 @@ class MessageReceiver:
         for consumer in self.messageConsumers:
             
             consumer.notify(message)
+
+
+# Receives messages
+def __demo_receive(port):
+
+    # Initialize the message receiver service
+    receiver = MessageReceiver(port)
+    
+    # Declare message consumers
+    class Consumer(implements(MessageConsumer)):
+
+        def notify(self, message): 
+
+            print(message)
+
+    consumer1 = Consumer()
+    consumer2 = Consumer()
+
+    # Register the consumers
+    receiver.registerConsumer(consumer1)
+    receiver.registerConsumer(consumer2)
+
+# Sends messages
+def __demo_send(sleepTime, port, host='127.0.0.1'):
+    
+    # Initialize sender
+    sender = MessageSender(host, port)
+    
+    # Send message
+    counter = 0
+    
+    while True:
+        
+        time.sleep(sleepTime)
+
+        sender.sendMessage("Counter: " + str(counter))
+        counter += 1
+
+if __name__ == '__main__':
+    port = 8000
+
+    # Start sender and receiver on two separate threads
+    threading.Thread(target=__demo_send, args= (1, port,)).start()
+    threading.Thread(target=__demo_receive, args = (port,)).start()
