--- plebnet/messaging/MessageConsumer.py
@@ -1,8 +1,0 @@
-from abc import ABC
-
-class MessageConsumer(ABC):
-
-    def notify(self, message):
-        
-        pass
-
