--- plebnet/messaging/__demo-receiver.py
@@ -1,3 +1,0 @@
-from __demo import receive
-
-receive(8000)
