--- plebnet/messaging/__demo-sender.py
@@ -1,3 +1,0 @@
-from __demo import send
-
-send(1, 8000)
