--- plebnet/messaging/__init__.py
@@ -1,3 +1,0 @@
-from plebnet.messaging.MessageSender import MessageSender
-from plebnet.messaging.MessageConsumer import MessageConsumer
-from plebnet.messaging.MessageReceiver import MessageReceiver
